% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{data.simulation}
\alias{data.simulation}
\title{Simulates data to use with odtr or sodtr}
\usage{
data.simulation(n, dA = NULL, a = NULL, kappa = NULL)
}
\arguments{
\item{n}{size of simulated data}

\item{kappa}{optional numeric indicating proportion of people who can receive treatment}
}
\description{
simulate data involving treatment, covariates and outcome for optimal or
simple dynamic treatment
}
