% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odtr.R
\name{odtr}
\alias{odtr}
\title{Optimal Dynamic Treatment Rules for Multiple Timepoints}
\usage{
odtr(
  data,
  Npsem,
  V,
  g_learner = "glm",
  Q_learner = "glm",
  type = c("binomial", "continuous"),
  maximize = TRUE
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A \code{data.frame} containing all necessary variables
for the estimation problem.}

\item{Npsem}{[\code{R6(Npsem)}]\cr
An \code{Npsem} object}

\item{V}{[\code{integer}]\cr
Number of folds for cross-fitting}

\item{g_learner}{[\code{character}]\cr A vector of \code{mlr3superlearner} algorithms
for estimation of the propensity score. Default is \code{"glm"}, a main effects GLM.}

\item{Q_learner}{[\code{character}]\cr A vector of \code{mlr3superlearner} algorithms
for estimation of the outcome regression. Default is \code{"glm"}, a main effects GLM.}

\item{type}{[\code{character(1)}]\cr
Outcome variable type (i.e., continuous, binomial).}

\item{maximize}{[\code{logical(1)}]\cr
Should the optimal treatment maximize (or minimize) the outcome?}
}
\value{
A \code{list} containing optimal treatment assignment and the estimated mean value of
the outcome under the optimal treatment rule.
}
\description{
Optimal Dynamic Treatment Rules for Multiple Timepoints
}
\examples{
dat <- odtr_sim(1e3)
sem <- Npsem$new(paste0("W", 1:4), A = c("A1", "A2"), Y = "Y")
optimal <- odtr(dat, sem, 1, "glm", "glm", "binomial")
}
