% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_rule.R
\name{optimal_rule}
\alias{optimal_rule}
\title{Optimal Dynamic Treatment Rules with Time-Varying Data}
\usage{
optimal_rule(
  data,
  Vars,
  folds,
  g_learner = "glm",
  Q_learner = "glm",
  type = c("binomial", "continuous"),
  maximize = TRUE
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A \code{data.frame} containing all necessary variables
for the estimation problem.}

\item{Vars}{[\code{R6(Vars)}]\cr
An \code{Vars} object}

\item{folds}{[\code{integer}]\cr
Number of folds for cross-fitting}

\item{g_learner}{[\code{character}]\cr
A \code{SuperLearner} library}

\item{Q_learner}{[\code{character}]\cr
A \code{SuperLearner} library}

\item{type}{[\code{character(1)}]\cr
Outcome variable type (i.e., continuous, binomial).}

\item{maximize}{[\code{logical(1)}]\cr
Should the optimal treatment maximize (or minimize) the outcome?}
}
\value{
A \code{list} containing optimal treatment assignment and the estimated mean value of
the outcome under the optimal treatment rule.
}
\description{
Optimal Dynamic Treatment Rules with Time-Varying Data
}
