% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SL.blip}
\alias{SL.blip}
\title{Additional Functions
Helps odtr to design optimal treatment rule and compute mean outcome}
\usage{
SL.blip(V, W, A, Y, QAW.reg, gAW, QAV.SL.library, risk.type, family, grid.size)
}
\description{
Additional Functions
Helps odtr to design optimal treatment rule and compute mean outcome
}
