% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SL.correct}
\alias{SL.correct}
\title{Additional Functions}
\usage{
SL.correct(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\description{
Helps odtr to design optimal treatment rule and compute mean outcome
}
