% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodtr.R
\name{sodtr}
\alias{sodtr}
\title{Compute Simple Dynamic Treatment Rule}
\usage{
sodtr(W, W_for_g, A, a, Y, rule, QAW.SL.library)
}
\arguments{
\item{W}{a dataframe of covariates}

\item{W_for_g}{a subset of W used to estimate the treatment mechanism}

\item{A}{a binary vector indicating observed treatment}

\item{a}{treatment for comparison with A}

\item{Y}{vector for outcome}

\item{rule}{vector of treatment (A) under dynamic treatment rule}

\item{QAW.SL.library}{Super learner library for estimating outcome regression}
}
\description{
Designs simple treatment rule and computes mean outcome
under this simple rule
}
