% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.fun.R
\name{tmle.fun}
\alias{tmle.fun}
\title{Computes TMLE Estimate of Mean Under Decision Rule}
\usage{
tmle.fun(A, Y, d, Qd, gAW, family)
}
\arguments{
\item{A}{a binary vector indicating observed treatment}

\item{Y}{vector for outcome}

\item{d}{output from dopt.fun}

\item{Qd}{Predicted outcome under optimal treatment}

\item{gAW}{estimate for probability of observed exposure}

\item{family}{character indicating family for glm
options: gaussian or binomial}
}
\description{
Helps odtr to design optimal treatment rule and compute mean outcome
}
