% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.simulation.R
\name{data.simulation}
\alias{data.simulation}
\title{Simulates Data for odtr or sodtr}
\usage{
data.simulation(n, dA = NULL, a = NULL, kappa = NULL)
}
\arguments{
\item{n}{size of simulated data}

\item{dA}{optional character indicating how to construct rule:
simple dynamic, ODTR or ODTR-RC
if ODTR-RC you must include kappa for constraint}

\item{a}{optional treatment for comparison with A}

\item{kappa}{optional numeric indicating proportion of people who can receive treatment}
}
\description{
simulate data involving treatment, covariates and outcome for optimal or
simple dynamic treatment
}
