% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopt.fun.R
\name{dopt.fun}
\alias{dopt.fun}
\title{Outputs Optimal d for tmle.fun}
\usage{
dopt.fun(blip, kappa)
}
\arguments{
\item{blip}{predicted blip values from SL.blip}

\item{kappa}{optional numeric indicating proportion of people who can receive treatment}
}
\description{
Helps odtr to design optimal treatment rule and compute mean outcome
}
