% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odtr.R
\name{odtr}
\alias{odtr}
\title{Optimal Dynamic Treatment Rules with Time-Varying Data}
\usage{
odtr(
  data,
  trt,
  outcome,
  baseline,
  time_varying,
  learners_trt = "glm",
  learners_outcome = "glm",
  learners_rule = "glm",
  folds,
  outcome_type = c("binomial", "continuous"),
  maximize = TRUE
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A \code{data.frame} containing all necessary variables
for the estimation problem.}

\item{trt}{}

\item{outcome}{}

\item{baseline}{}

\item{learners_trt}{[\code{character}]\cr
A \code{mlr3superlearner} library}

\item{learners_outcome}{[\code{character}]\cr
A \code{mlr3superlearner} library}

\item{learners_rule}{[\code{character}]\cr
A \code{mlr3superlearner} library}

\item{folds}{[\code{integer}]\cr
Number of folds for cross-fitting}

\item{outcome_type}{[\code{character(1)}]\cr
Outcome variable type (i.e., continuous, binomial).}

\item{maximize}{[\code{logical(1)}]\cr
Should the optimal treatment maximize (or minimize) the outcome?}
}
\value{
A \code{list} containing optimal treatment assignment and the estimated mean value of the outcome under the optimal treatment rule.
}
\description{
Optimal Dynamic Treatment Rules with Time-Varying Data
}
