% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odtr.R
\name{odtr}
\alias{odtr}
\title{Compute Optimal Dynamic Treatment Rule}
\usage{
odtr(W, W_for_g, A, a, Y, V, QAW.SL.library, QAV.SL.library,
boundsY = c(0,1), risk.type="empirical", grid.size = 100, kappa = NULL)
}
\arguments{
\item{W}{a dataframe of covariates}

\item{W_for_g}{a subset of W used to estimate the treatment mechanism}

\item{A}{a binary vector indicating observed treatment}

\item{a}{treatment for comparison with A}

\item{Y}{vector for outcome}

\item{V}{subset of W used to construct optimal rule}

\item{QAW.SL.library}{Super learner library for estimating outcome regression}

\item{QAV.SL.library}{Super learner library for estimating blip function}

\item{boundsY}{bounds for outcome variable Y default: c(0,1)}

\item{risk.type}{character of length 1 indicating
risk to be minimized when estimating optimal treatment can be: empirical, TMLE, CV empirical, CV TMLE
default empirical}

\item{grid.size}{numeric of length 1 indicating size of grid on which to do grid search
for weights assigned to algorithms used to estimate optimal treatment
default 100}

\item{kappa}{(optional) numeric indicating minimum proportion of individuals who can receive treatment (A=1)}
}
\description{
Designs optimal treatment rule and computes mean outcome
under this optimal rule
}
