% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.blip.R
\name{SL.blip}
\alias{SL.blip}
\title{Computes Blips and Updates SuperLearner and Optimal Rule}
\usage{
SL.blip(V, W, A, Y, QAW.reg, gAW, QAV.SL.library, risk.type, family, grid.size)
}
\arguments{
\item{V}{subset of W used to construct optimal rule}

\item{W}{a dataframe of covariates}

\item{A}{a binary vector indicating observed treatment}

\item{Y}{vector for outcome}

\item{QAW.reg}{Output from SuperLearner E{Y|A,W}}

\item{gAW}{estimate for probability of observed exposure}

\item{QAV.SL.library}{Super learner library for estimating blip function}

\item{risk.type}{character of length 1 indicating
risk to be minimized when estimating optimal treatment can be: empirical, TMLE, CV empirical, CV TMLE
default empirical}

\item{family}{character indicating family for glm
options: gaussian or binomial}

\item{grid.size}{numeric of length 1 indicating size of grid on which to do grid search
for weights assigned to algorithms used to estimate optimal treatment
default 100}
}
\description{
Helps odtr to design optimal treatment rule and compute mean outcome
}
