% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.correct.R
\name{SL.correct}
\alias{SL.correct}
\title{Checks for Correctly Specified QAW Regression}
\usage{
SL.correct(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\arguments{
\item{Y}{vector for outcome varible}

\item{X}{dataframe for response variables, could be a matrix}

\item{newX}{dataframe with new response variables for prediction}

\item{family}{family for glm fit for options see glm}

\item{obsWeights}{optional weihgts to be applied to glm}

\item{model}{logic indicating model frame should be included as output default TRUE}

\item{...}{additional arguments to be passed to glm}
}
\description{
Helps odtr to design optimal treatment rule and compute mean outcome
}
